+++ src/definition/array-function.js
@@ -1,8 +1,0 @@
-module.exports = {
-  'array-func/from-map': 2,
-  'array-func/no-unnecessary-this-arg': 2,
-  'array-func/prefer-array-from': 2,
-  'array-func/avoid-reverse': 2,
-  'array-func/prefer-flat-map': 1,
-  'array-func/prefer-flat': 1
-}