+++ src/definition/filenames.js
@@ -1,5 +1,0 @@
-module.exports = {
-  'filenames/match-exported': [2, [null, 'kebab', 'pascal', 'camel']],
-  'filenames/match-regex': 0,
-  'filenames/no-index': 0
-}