+++ src/definition/you-dont-need-lodash-underscore.js
@@ -1,5 +1,5 @@
 module.exports = {
-  'you-dont-need-lodash-underscore/for-each': 1,
-  'you-dont-need-lodash-underscore/concat': 1,
-  'you-dont-need-lodash-underscore/map': 1
+  'you-dont-need-lodash-underscore/concat': 'warn',
+  'you-dont-need-lodash-underscore/for-each': 'warn',
+  'you-dont-need-lodash-underscore/map': 'warn'
 }