+++ src/definition/you-dont-need-momentjs.js
@@ -1,6 +1,6 @@
 module.exports = {
-  'you-dont-need-momentjs/no-dynamic-import-moment': 2,
-  'you-dont-need-momentjs/no-import-moment': 2,
-  'you-dont-need-momentjs/no-moment-constructor': 2,
-  'you-dont-need-momentjs/no-require-moment': 2
+  'you-dont-need-momentjs/no-dynamic-import-moment': 'error',
+  'you-dont-need-momentjs/no-import-moment': 'error',
+  'you-dont-need-momentjs/no-moment-constructor': 'error',
+  'you-dont-need-momentjs/no-require-moment': 'error'
 }