+++ src/plugins.js
@@ -1,14 +1,20 @@
-module.exports = [
-  'array-func',
+const isInstalled = require('is-installed')
+
+const plugins = [
   'babel',
-  'filenames',
+  'html',
   'import',
-  'jest',
   'json',
   'node',
   'promise',
   'security',
   'unicorn',
-  'you-dont-need-lodash-underscore',
-  'you-dont-need-momentjs'
 ]
+
+if (isInstalled.sync('jest')) plugins.push('jest')
+
+if (isInstalled.sync('lodash')) plugins.push('you-dont-need-lodash-underscore')
+
+if (isInstalled.sync('momentjs')) plugins.push('you-dont-need-momentjs')
+
+module.exports = plugins