+++ src/settings.js
@@ -1,0 +1,4 @@
+module.exports = {
+  'html/indent': '+2',
+  'html/report-bad-indent': 'warn'
+}